+++ app/ExternalFiles/Forms.php
@@ -91,6 +91,9 @@ class Forms {
 		add_action( 'eml_before_file_list', array( $this, 'set_import_max' ), 10, 2 );
 		add_filter( 'eml_import_urls', array( $this, 'filter_urls' ) );
 		add_action( 'eml_after_file_save', array( $this, 'add_imported_url_to_list' ), 10, 3 );
+
+		// misc.
+		add_filter( 'admin_body_class', array( $this, 'add_sound' ) );
 	}
 
 	/**
@@ -197,6 +200,8 @@ class Forms {
 				'title_sync_progress'           => __( 'Synchronization in progress', 'external-files-in-media-library' ),
 				'title_loading'                 => __( 'Loading ..', 'external-files-in-media-library' ),
 				'text_loading'                  => __( 'Please wait a moment ..', 'external-files-in-media-library' ),
+				/* source of file: https://pixabay.com */
+				'success_sound_file'            => Helper::get_plugin_url() . 'gfx/success.mp3'
 			)
 		);
 	}
@@ -928,5 +933,25 @@ class Forms {
 
 		// add the error object to the list of errors.
 		Results::get_instance()->add( $result_obj );
+	}
+
+	/**
+	 * Enabled to play a sound if import finishes.
+	 *
+	 * @param string $classes List of classes as string.
+	 *
+	 * @return string
+	 */
+	public function add_sound( string $classes ): string {
+		// bail if setting is not enabled.
+		if( 1 !== absint( get_option( 'eml_play_sound' ) ) ) {
+			return $classes;
+		}
+
+		// add the class.
+		$classes .= ' efml-play-found';
+
+		// return resulting list of classes.
+		return $classes;
 	}
 }